Handle UDP bind -> listen EACCES.
`IO::Endpoint::Wrapper.bind` always calls `Socket#listen`,
and handles `Errno::EOPNOTSUPP`, for example for UDP sockets,
for which the `listen` call is not applicable (see `man 2 listen`).
However, in some environments, such as in a container,
the error may be mapped to `EACCES`,
which can be verified with a simple C program.
This patch ignores `Errno::EACCES` if it was a UDP socket,
since the `listen` call was meaningless in the first place.